+++ tests/call_fn.rs
@@ -116,7 +116,7 @@ fn test_anonymous_fn() -> Result<(), Box<EvalAltResult>> {
 
 #[test]
 #[cfg(not(feature = "no_object"))]
-fn test_fn_ptr() -> Result<(), Box<EvalAltResult>> {
+fn test_fn_ptr_raw() -> Result<(), Box<EvalAltResult>> {
     let mut engine = Engine::new();
 
     engine.register_raw_fn(